renamed context 'attribution' to 'effective_class_repartition'
