+++ ddd/logic/effective_class_repartition/builder/tutor_identity_builder.py
@@ -23,27 +23,19 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from typing import Optional
-
-from ddd.logic.effective_class_repartition.commands import SearchTutorsDistributedToClassCommand
-from ddd.logic.learning_unit.domain.model.effective_class import EffectiveClassIdentity
-from ddd.logic.learning_unit.domain.service.i_tutor_assigned_to_class import ITutorAssignedToClassTranslator
+from ddd.logic.effective_class_repartition.domain.model.tutor import TutorIdentity
+from osis_common.ddd.interface import EntityIdentityBuilder, DTO
 
 
-class TutorAssignedToClassTranslator(ITutorAssignedToClassTranslator):
+class TutorIdentityBuilder(EntityIdentityBuilder):
 
     @classmethod
-    def get_first_tutor_full_name_if_exists(
+    def build_from_repository_dto(cls, dto_object: 'DTO') -> 'TutorIdentity':
+        raise NotImplementedError
+
+    @classmethod
+    def build_from_personal_id_number(
             cls,
-            effective_class_identity: 'EffectiveClassIdentity'
-    ) -> Optional[str]:
-        from infrastructure.messages_bus import message_bus_instance
-        tutors_assigned_to_class = message_bus_instance.invoke(
-            SearchTutorsDistributedToClassCommand(
-                class_code=effective_class_identity.class_code,
-                learning_unit_code=effective_class_identity.learning_unit_identity.code,
-                learning_unit_year=effective_class_identity.learning_unit_identity.year
-            )
-        )
-        if tutors_assigned_to_class:
-            return tutors_assigned_to_class[0].full_name
+            personal_id_number: str
+    ) -> 'TutorIdentity':
+        return TutorIdentity(personal_id_number=personal_id_number)
