+++ ddd/logic/effective_class_repartition/domain/model/_class_volume_repartition.py
@@ -23,27 +23,23 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from typing import Optional
+from decimal import Decimal
 
-from ddd.logic.effective_class_repartition.commands import SearchTutorsDistributedToClassCommand
+import attr
+
+from ddd.logic.effective_class_repartition.domain.model._learning_unit_attribution import LearningUnitAttributionIdentity
 from ddd.logic.learning_unit.domain.model.effective_class import EffectiveClassIdentity
-from ddd.logic.learning_unit.domain.service.i_tutor_assigned_to_class import ITutorAssignedToClassTranslator
+from osis_common.ddd import interface
 
 
-class TutorAssignedToClassTranslator(ITutorAssignedToClassTranslator):
+@attr.s(slots=True, hash=False, eq=False)
+class ClassVolumeRepartition(interface.ValueObject):
+    effective_class = attr.ib(type=EffectiveClassIdentity)
+    distributed_volume = attr.ib(type=Decimal)
+    attribution = attr.ib(type=LearningUnitAttributionIdentity)
 
-    @classmethod
-    def get_first_tutor_full_name_if_exists(
-            cls,
-            effective_class_identity: 'EffectiveClassIdentity'
-    ) -> Optional[str]:
-        from infrastructure.messages_bus import message_bus_instance
-        tutors_assigned_to_class = message_bus_instance.invoke(
-            SearchTutorsDistributedToClassCommand(
-                class_code=effective_class_identity.class_code,
-                learning_unit_code=effective_class_identity.learning_unit_identity.code,
-                learning_unit_year=effective_class_identity.learning_unit_identity.year
-            )
-        )
-        if tutors_assigned_to_class:
-            return tutors_assigned_to_class[0].full_name
+    def __hash__(self):
+        return hash("{}{}".format(self.effective_class, self.attribution))
+
+    def __eq__(self, other):
+        return hash(self) == hash(other)
