+++ ddd/logic/effective_class_repartition/domain/model/_learning_unit_attribution.py
