+++ ddd/logic/learning_unit/builder/effective_class_identity_builder.py
@@ -25,7 +25,7 @@
 ##############################################################################
 from typing import Union
 
-from ddd.logic.attribution.commands import DistributeClassToTutorCommand
+from ddd.logic.effective_class_repartition.commands import DistributeClassToTutorCommand
 from ddd.logic.learning_unit.builder.learning_unit_identity_builder import LearningUnitIdentityBuilder
 from ddd.logic.learning_unit.commands import CreateEffectiveClassCommand, UpdateEffectiveClassCommand, \
     DeleteEffectiveClassCommand
